--- a/.github/CODEOWNERS
+++ b/.github/CODEOWNERS
@@ -1,0 +1,3 @@
+# Identify the committers responsible for the repository as standard reviewers
+
+* @rafaelmag110 @ndr-brt @lgblaumeiser