--- a/.github/workflows/secrets-scan.yml
+++ b/.github/workflows/secrets-scan.yml
@@ -46,7 +46,7 @@ jobs:
 
       - name: TruffleHog OSS
         id: trufflehog
-        uses: trufflesecurity/trufflehog@a2a17cd73d74376209d6323c80a9a55b424e25b0
+        uses: trufflesecurity/trufflehog@ca270a7e14a3542229c2cadf1d408ecac1455815
         continue-on-error: true
         with:
           path: ./  # Scan the entire repository