--- a/.github/workflows/secrets-scan.yml
+++ b/.github/workflows/secrets-scan.yml
@@ -46,7 +46,7 @@ jobs:
 
       - name: TruffleHog OSS
         id: trufflehog
-        uses: trufflesecurity/trufflehog@ca270a7e14a3542229c2cadf1d408ecac1455815
+        uses: trufflesecurity/trufflehog@12164e38f0f1b673ab0594c7d94daf71b0be6823
         continue-on-error: true
         with:
           path: ./  # Scan the entire repository