--- a/.tractusx
+++ b/.tractusx
@@ -2,4 +2,5 @@ product: "Tractus-X EDC"
 leadingRepository: "https://github.com/eclipse-tractusx/tractusx-edc"
 repositories: []
 openApiSpecs:
-- "https://api.swaggerhub.com/apis/eclipse-tractusx-bot/tractusx-edc/0.7.3/swagger.yaml"
+- "https://eclipse-tractusx.github.io/tractusx-edc/openapi/control-plane-api/control-plane.yaml"
+- "https://eclipse-tractusx.github.io/tractusx-edc/openapi/data-plane-api/data-plane.yaml"