--- a/build.gradle.kts
+++ b/build.gradle.kts
@@ -75,7 +75,7 @@ allprojects {
             implementation("com.azure:azure-core-http-netty:1.15.10") {
                 because("Depends on netty-handler:4.1.115.Final that has a vunlnerability: https://ossindex.sonatype.org/component/pkg:maven/io.netty/netty-handler@4.1.115.Final")
             }
-            implementation("software.amazon.awssdk:netty-nio-client:2.30.17") {
+            implementation("software.amazon.awssdk:netty-nio-client:2.30.36") {
                 because("Depends on netty-handler:4.1.115.Final that has a vunlnerability: https://ossindex.sonatype.org/component/pkg:maven/io.netty/netty-handler@4.1.115.Final")
             }
         }