--- a/gradle.properties
+++ b/gradle.properties
@@ -1,5 +1,5 @@
 group=org.eclipse.tractusx.edc
-version=0.9.0-SNAPSHOT
+version=0.9.0
 # configure the build:
 txScmConnection=scm:git:git@github.com:eclipse-tractusx/tractusx-edc.git
 txWebsiteUrl=https://github.com/eclipse-tractusx/tractusx-edc.git