--- a/gradle.properties
+++ b/gradle.properties
@@ -4,3 +4,5 @@ version=0.9.0
 txScmConnection=scm:git:git@github.com:eclipse-tractusx/tractusx-edc.git
 txWebsiteUrl=https://github.com/eclipse-tractusx/tractusx-edc.git
 txScmUrl=https://github.com/eclipse-tractusx/tractusx-edc.git
+
+org.gradle.caching=true